requirements.txt
Flask==2.1.2
mysql==0.0.3
redis==4.2.2
